Refactor transcription handlers
