+++ modules/transcription-service-workflowoperation/src/main/java/org/opencastproject/handler/workflowoperation/AmberscriptStartTranscriptionOperationHandler.java
@@ -18,8 +18,7 @@
  * the License.
  *
  */
-
-package org.opencastproject.transcription.workflowoperation;
+package org.opencastproject.handler.workflowoperation;
 
 import org.opencastproject.job.api.Job;
 import org.opencastproject.job.api.JobContext;
@@ -53,23 +52,26 @@ import java.util.Collection;
 import java.util.List;
 
 @Component(
-        immediate = true,
-        service = WorkflowOperationHandler.class,
-        property = {
-                "service.description=Microsoft Azure Start Transcription Workflow Operation Handler",
-                "workflow.operation=microsoft-azure-start-transcription"
-        }
+    immediate = true,
+    service = WorkflowOperationHandler.class,
+    property = {
+        "service.description=Start Transcription Workflow Operation Handler (Amberscript)",
+        "workflow.operation=amberscript-start-transcription"
+    }
 )
-public class MicrosoftAzureStartTranscriptionOperationHandler extends AbstractWorkflowOperationHandler {
+public class AmberscriptStartTranscriptionOperationHandler extends AbstractWorkflowOperationHandler {
 
-  /** The logging facility */
-  private static final Logger logger = LoggerFactory.getLogger(MicrosoftAzureStartTranscriptionOperationHandler.class);
+  private static final Logger logger = LoggerFactory.getLogger(AmberscriptStartTranscriptionOperationHandler.class);
 
   /** Workflow configuration option keys */
-  static final String OPT_LANGUAGE_CODE = "language-code";
-  static final String OPT_SKIP_IF_FLAVOR_EXISTS = "skip-if-flavor-exists";
-  static final String OPT_AUTO_DETECT_LANGUAGE = "auto-detect-language";
-  static final String OPT_AUTO_DETECT_LANGUAGES = "auto-detect-languages";
+  static final String LANGUAGE = "language";
+  static final String JOBTYPE = "jobtype";
+  static final String SPEAKER = "speaker";
+  static final String TRANSCRIPTIONTYPE = "transcriptiontype";
+  static final String GLOSSARY = "glossary";
+  static final String TRANSCRIPTIONSTYLE = "transcriptionstyle";
+  static final String TARGETLANGUAGE = "targetlanguage";
+  static final String SKIP_IF_FLAVOR_EXISTS = "skip-if-flavor-exists";
 
   /** The transcription service */
   private TranscriptionService service = null;
@@ -86,24 +88,35 @@ public class MicrosoftAzureStartTranscriptionOperationHandler extends AbstractWo
     MediaPackage mediaPackage = workflowInstance.getMediaPackage();
     WorkflowOperationInstance operation = workflowInstance.getCurrentOperation();
 
-    String skipOption = StringUtils.trimToNull(operation.getConfiguration(OPT_SKIP_IF_FLAVOR_EXISTS));
+    String skipOption = StringUtils.trimToNull(operation.getConfiguration(SKIP_IF_FLAVOR_EXISTS));
     if (skipOption != null) {
       MediaPackageElement[] mpes = mediaPackage.getElementsByFlavor(MediaPackageElementFlavor.parseFlavor(skipOption));
       if (mpes != null && mpes.length > 0) {
         logger.info(
-                "Start transcription operation will be skipped because flavor {} already exists in the media package",
-                skipOption);
+            "Start transcription operation will be skipped because flavor '{}' already exists in the media package.",
+            skipOption);
         return createResult(Action.SKIP);
       }
     }
 
-    logger.debug("Start transcription for mediapackage {} started", mediaPackage);
+    logger.debug("Start transcription for mediapackage '{}'.", mediaPackage);
 
     // Check which tags have been configured
     ConfiguredTagsAndFlavors tagsAndFlavors = getTagsAndFlavors(
-            workflowInstance, Configuration.many, Configuration.many, Configuration.none, Configuration.none);
+        workflowInstance, Configuration.many, Configuration.many, Configuration.none, Configuration.none);
     List<String> sourceTagOption = tagsAndFlavors.getSrcTags();
     List<MediaPackageElementFlavor> sourceFlavorOption = tagsAndFlavors.getSrcFlavors();
+    String language = StringUtils.trimToEmpty(operation.getConfiguration(LANGUAGE));
+    String jobType = StringUtils.trimToEmpty(operation.getConfiguration(JOBTYPE));
+    String speaker = StringUtils.trimToEmpty(operation.getConfiguration(SPEAKER));
+    String transcriptionType = StringUtils.trimToEmpty(operation.getConfiguration(TRANSCRIPTIONTYPE));
+    // Note that specifying `""` is different from not specifying a glossary at all!
+    // The former will override the default with not using a glossary for this operation,
+    // while the latter will fall back to said default. Hence, no `trimToEmpty` here.
+    String glossary = StringUtils.trim(operation.getConfiguration(GLOSSARY));
+    String transcriptionStyle = StringUtils.trimToEmpty(operation.getConfiguration(TRANSCRIPTIONSTYLE));
+    // See glossary comment above
+    String targetLanguage = StringUtils.trim(operation.getConfiguration(TARGETLANGUAGE));
 
     AbstractMediaPackageElementSelector<Track> elementSelector = new TrackSelector();
 
@@ -113,33 +126,18 @@ public class MicrosoftAzureStartTranscriptionOperationHandler extends AbstractWo
     }
 
     if (!sourceFlavorOption.isEmpty()) {
-      MediaPackageElementFlavor flavor = sourceFlavorOption.get(0);
-      elementSelector.addFlavor(flavor);
+      elementSelector.addFlavor(sourceFlavorOption.get(0));
     }
     if (!sourceTagOption.isEmpty()) {
       elementSelector.addTag(sourceTagOption.get(0));
-    }
-
-    // Get language code if configured
-    String langCode = operation.getConfiguration(OPT_LANGUAGE_CODE);
-    if (StringUtils.isNotBlank(langCode)) {
-      langCode = StringUtils.trim(langCode);
-    }
-    String autoDetectLanguage = operation.getConfiguration(OPT_AUTO_DETECT_LANGUAGE);
-    if (StringUtils.isNotBlank(langCode)) {
-      langCode = StringUtils.trim(langCode);
-    }
-    String autoDetectLanguages = operation.getConfiguration(OPT_AUTO_DETECT_LANGUAGES);
-    if (StringUtils.isNotBlank(langCode)) {
-      langCode = StringUtils.trim(langCode);
     }
 
     Collection<Track> elements = elementSelector.select(mediaPackage, false);
     Job job = null;
     for (Track track : elements) {
       try {
-        job = service.startTranscription(mediaPackage.getIdentifier().toString(), track, langCode, autoDetectLanguage,
-                autoDetectLanguages);
+        job = service.startTranscription(mediaPackage.getIdentifier().toString(), track, language, jobType, speaker,
+            transcriptionType, glossary, transcriptionStyle, targetLanguage);
         // Only one job per media package
         break;
       } catch (TranscriptionServiceException e) {
@@ -148,23 +146,23 @@ public class MicrosoftAzureStartTranscriptionOperationHandler extends AbstractWo
     }
 
     if (job == null) {
-      logger.info("No matching tracks found");
+      logger.info("No matching tracks found.");
       return createResult(mediaPackage, Action.CONTINUE);
     }
 
     // Wait for the jobs to return
     if (!waitForStatus(job).isSuccess()) {
-      throw new WorkflowOperationException("Transcription job did not complete successfully");
+      throw new WorkflowOperationException("Transcription job did not complete successfully.");
     }
-    // Return OK means that the microsoft azure job was created, but not finished yet
+    // Return OK means that the transcription job was created, but not finished yet
 
-    logger.debug("External transcription job for mediapackage {} was created", mediaPackage);
+    logger.debug("External transcription job for mediapackage '{}' was created.", mediaPackage);
 
     // Results are empty, we should get a callback when transcription is done
     return createResult(Action.CONTINUE);
   }
 
-  @Reference(target = "(provider=microsoft.azure)")
+  @Reference(target = "(provider=amberscript)")
   public void setTranscriptionService(TranscriptionService service) {
     this.service = service;
   }
