+++ modules/transcription-service-workflowoperation/src/main/java/org/opencastproject/handler/workflowoperation/NibityAttachTranscriptionOperationHandler.java
@@ -18,7 +18,7 @@
  * the License.
  *
  */
-package org.opencastproject.transcription.workflowoperation;
+package org.opencastproject.handler.workflowoperation;
 
 import org.opencastproject.job.api.JobContext;
 import org.opencastproject.mediapackage.Attachment;
