+++ modules/transcription-service-workflowoperation/src/test/java/org/opencastproject/handler/workflowoperation/GoogleSpeechAttachTranscriptionOperationHandlerTest.java
@@ -18,7 +18,7 @@
  * the License.
  *
  */
-package org.opencastproject.transcription.workflowoperation;
+package org.opencastproject.handler.workflowoperation;
 
 import org.opencastproject.caption.api.CaptionService;
 import org.opencastproject.job.api.Job;
