+++ modules/transcription-service-workflowoperation/src/test/java/org/opencastproject/handler/workflowoperation/MicrosoftAzureAttachTranscriptionOperationHandlerTest.java
@@ -18,7 +18,7 @@
  * the License.
  *
  */
-package org.opencastproject.transcription.workflowoperation;
+package org.opencastproject.handler.workflowoperation;
 
 import org.opencastproject.job.api.Job;
 import org.opencastproject.mediapackage.Attachment;
