+++ modules/transcription-service-workflowoperation/src/test/java/org/opencastproject/handler/workflowoperation/MicrosoftAzureStartTranscriptionOperationHandlerTest.java
@@ -18,7 +18,8 @@
  * the License.
  *
  */
-package org.opencastproject.transcription.workflowoperation;
+
+package org.opencastproject.handler.workflowoperation;
 
 import org.opencastproject.job.api.Job;
 import org.opencastproject.mediapackage.MediaPackage;
@@ -31,7 +32,6 @@ import org.opencastproject.workflow.api.WorkflowDefinitionImpl;
 import org.opencastproject.workflow.api.WorkflowInstance;
 import org.opencastproject.workflow.api.WorkflowOperationException;
 import org.opencastproject.workflow.api.WorkflowOperationInstance;
-import org.opencastproject.workflow.api.WorkflowOperationInstance.OperationState;
 import org.opencastproject.workflow.api.WorkflowOperationResult;
 import org.opencastproject.workflow.api.WorkflowOperationResult.Action;
 
@@ -46,15 +46,21 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-public class StartTranscriptionOperationHandlerTest {
+public class MicrosoftAzureStartTranscriptionOperationHandlerTest {
 
-  /** The operation handler to test */
-  private StartTranscriptionOperationHandler operationHandler;
+  /**
+   * The operation handler to test
+   */
+  private MicrosoftAzureStartTranscriptionOperationHandler operationHandler;
 
-  /** The transcription service */
+  /**
+   * The transcription service
+   */
   private TranscriptionService service;
 
-  /** The operation instance */
+  /**
+   * The operation instance
+   */
   private WorkflowOperationInstance operation;
 
   private MediaPackage mediaPackage;
@@ -66,7 +72,8 @@ public class StartTranscriptionOperationHandlerTest {
     MediaPackageBuilder builder = MediaPackageBuilderFactory.newInstance().newMediaPackageBuilder();
 
     // Media package set up
-    URI mediaPackageURI = StartTranscriptionOperationHandlerTest.class.getResource("/mp.xml").toURI();
+    URI mediaPackageURI = MicrosoftAzureStartTranscriptionOperationHandlerTest.class.getResource("/mp_google.xml")
+            .toURI();
     mediaPackage = builder.loadFromXml(mediaPackageURI.toURL().openStream());
 
     // Service registry set up
@@ -76,7 +83,7 @@ public class StartTranscriptionOperationHandlerTest {
     EasyMock.expect(job1.getStatus()).andReturn(Job.Status.FINISHED);
     EasyMock.expect(job1.getDateCreated()).andReturn(new Date());
     EasyMock.expect(job1.getDateStarted()).andReturn(new Date());
-    EasyMock.expect(job1.getQueueTime()).andReturn(new Long(0));
+    EasyMock.expect(job1.getQueueTime()).andReturn(0L);
     EasyMock.replay(job1);
 
     ServiceRegistry serviceRegistry = EasyMock.createNiceMock(ServiceRegistry.class);
@@ -86,29 +93,33 @@ public class StartTranscriptionOperationHandlerTest {
     // Transcription service set up
     service = EasyMock.createStrictMock(TranscriptionService.class);
     capturedTrack = Capture.newInstance();
-    EasyMock.expect(service.startTranscription(EasyMock.anyObject(String.class), EasyMock.capture(capturedTrack)))
+    EasyMock
+            .expect(service.startTranscription(
+                    EasyMock.anyObject(String.class), EasyMock.capture(capturedTrack), EasyMock.anyObject(String.class),
+                    EasyMock.anyObject(String.class), EasyMock.anyObject(String.class))
+            )
             .andReturn(null);
     EasyMock.replay(service);
 
     // Workflow set up
     WorkflowDefinitionImpl def = new WorkflowDefinitionImpl();
-    def.setId("DCE-start-transcription");
+    def.setId("microsoft-azure-start-transcription");
     workflowInstance = new WorkflowInstance(def, mediaPackage, null, null, null);
     workflowInstance.setId(1);
-    operation = new WorkflowOperationInstance("start-transcript", OperationState.RUNNING);
+    operation = new WorkflowOperationInstance("start-transcript", WorkflowOperationInstance.OperationState.RUNNING);
     List<WorkflowOperationInstance> operationList = new ArrayList<WorkflowOperationInstance>();
     operationList.add(operation);
     workflowInstance.setOperations(operationList);
 
     // Operation handler set up
-    operationHandler = new StartTranscriptionOperationHandler();
+    operationHandler = new MicrosoftAzureStartTranscriptionOperationHandler();
     operationHandler.setTranscriptionService(service);
     operationHandler.setServiceRegistry(serviceRegistry);
   }
 
   @Test
   public void testStartSelectByFlavor() throws Exception {
-    operation.setConfiguration(StartTranscriptionOperationHandler.SOURCE_FLAVOR, "audio/ogg");
+    operation.setConfiguration(MicrosoftAzureStartTranscriptionOperationHandler.SOURCE_FLAVOR, "audio/flac");
 
     WorkflowOperationResult result = operationHandler.start(workflowInstance, null);
     Assert.assertEquals(Action.CONTINUE, result.getAction());
@@ -118,10 +129,10 @@ public class StartTranscriptionOperationHandlerTest {
 
   @Test
   public void testStartSelectByTag() throws Exception {
-    operation.setConfiguration(StartTranscriptionOperationHandler.SOURCE_TAG, "transcript");
+    operation.setConfiguration(MicrosoftAzureStartTranscriptionOperationHandler.SOURCE_TAG, "transcript");
 
     WorkflowOperationResult result = operationHandler.start(workflowInstance, null);
-    Assert.assertEquals(Action.CONTINUE, result.getAction());
+    Assert.assertEquals(WorkflowOperationResult.Action.CONTINUE, result.getAction());
 
     Assert.assertEquals("audioTrack1", capturedTrack.getValue().getIdentifier());
   }
@@ -129,7 +140,8 @@ public class StartTranscriptionOperationHandlerTest {
   @Test
   public void testStartSkipFlavor() throws Exception {
     // Make sure operation will be skipped if media package already contains the flavor passed
-    operation.setConfiguration(StartTranscriptionOperationHandler.SKIP_IF_FLAVOR_EXISTS, "audio/ogg");
+    operation.setConfiguration(MicrosoftAzureStartTranscriptionOperationHandler.OPT_SKIP_IF_FLAVOR_EXISTS,
+            "audio/flac");
 
     WorkflowOperationResult result = operationHandler.start(workflowInstance, null);
     Assert.assertEquals(Action.SKIP, result.getAction());
@@ -137,9 +149,10 @@ public class StartTranscriptionOperationHandlerTest {
 
   @Test
   public void testStartDontSkipFlavor() throws Exception {
-    operation.setConfiguration(StartTranscriptionOperationHandler.SOURCE_TAG, "transcript");
+    operation.setConfiguration(MicrosoftAzureStartTranscriptionOperationHandler.SOURCE_TAG, "transcript");
     // Make sure operation will NOT be skipped if media package does NOT contain the flavor passed
-    operation.setConfiguration(StartTranscriptionOperationHandler.SKIP_IF_FLAVOR_EXISTS, "captions/timedtext");
+    operation.setConfiguration(
+            MicrosoftAzureStartTranscriptionOperationHandler.OPT_SKIP_IF_FLAVOR_EXISTS, "captions/timedtext");
 
     WorkflowOperationResult result = operationHandler.start(workflowInstance, null);
     Assert.assertEquals(Action.CONTINUE, result.getAction());
